requirements.txt
django
pillow
pdf2image
yt-dlp


